Fix review comments

+++ lily/deals/migrations/0033_auto_20161025_1415.py
@@ -34,7 +34,7 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='dealwhycustomer',
             name='position',
-            field=models.IntegerField(default=0),
+            field=models.PositiveSmallIntegerField(default=0),
         ),
         migrations.AlterField(
             model_name='dealwhylost',
